Source plugin config files for neovim HM setup
The neovim configuration file nvim.nix was getting too long. Further,
the lua code was specified by literal strings without context, making
it so parsers, code completion, etc. would not work. Separating out
the plugin specific configuration fixes both of these issues. It also
has the side benefit of making the neovim files more useful for non-nix
based configs. The only downside to this is that it relies on PR
submitted to home-manager that may or may not be accepted for merge.