Sets up base lspconfig setup for neovim
The configuration is left as a simple config for pyright only because
neovim on nixos unstable is still at v0.6.1. When v0.7.0+ comes to
nixos unstable, a simple config is ideal for testing functionality
with the new lua bindings.

Additionally, modularization of the development tools (e.g. python3
and pyright) belong on the user config level. It is possible that a
user config may not want a system install of these packages. Progress
has been started to shift development tools that were previously
system wide to home-manager under appropriate user modules.